+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Add `img` selector and matcher
 - Added `aria` filter to `:combo_box`, `:disclosure`, and `:disclosure_button` [Sean Doyle]
 - Removed support for Ruby 2.7. Minimum supported Ruby version is now 3.0
+- Added a ':dialog' selector
+- The `:modal` selector will now check a native modal is displaying as a modal
 
 ## v0.10.0